+++ api-augment/substrate/registry.ts
@@ -1,4 +1,4 @@
 // Copyright 2017-2022 @polkadot/api-augment authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import 'https://deno.land/x/polkadot@0.2.18/types-augment/registry/substrate.ts';
+import 'https://deno.land/x/polkadot/types-augment/registry/substrate.ts';