+++ api-contract/promise/index.ts
@@ -1,9 +1,9 @@
 // Copyright 2017-2022 @polkadot/api-contract authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { AccountId, Hash } from 'https://deno.land/x/polkadot@0.2.18/types/interfaces/index.ts';
+import type { AccountId, Hash } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
 
-import { ApiPromise, toPromiseMethod } from 'https://deno.land/x/polkadot@0.2.18/api/mod.ts';
+import { ApiPromise, toPromiseMethod } from 'https://deno.land/x/polkadot/api/mod.ts';
 
 import { Abi } from '../Abi/index.ts';
 import { Blueprint, Code, Contract } from '../base/index.ts';