+++ types/interfaces/babe/runtime.ts
@@ -3,7 +3,7 @@
 
 import type { DefinitionCall, DefinitionsCall } from '../../types/index.ts';
 
-import { objectSpread } from 'https://deno.land/x/polkadot@0.2.18/util/mod.ts';
+import { objectSpread } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 const V1_V2_SHARED: Record<string, DefinitionCall> = {
   current_epoch: {