+++ types/interfaces/gilt/types.ts
@@ -1,8 +1,8 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
-import type { Struct, u32 } from 'https://deno.land/x/polkadot@0.2.18/types-codec/mod.ts';
-import type { AccountId, Balance, BlockNumber, Perquintill } from 'https://deno.land/x/polkadot@0.2.18/types/interfaces/runtime/index.ts';
+import type { Struct, u32 } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
+import type { AccountId, Balance, BlockNumber, Perquintill } from 'https://deno.land/x/polkadot/types/interfaces/runtime/index.ts';
 
 /** @name ActiveGilt */
 export interface ActiveGilt extends Struct {